+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix a bug in Pipeline Management table where sometimes pipeline action buttons do not send requests ([#1008](https://github.com/feldera/feldera/pull/1008))
+- Fix pipeline with no program displays COMPILING status in Pipeline Management table
 
 ## [0.3.1] - 2023-11-09